pet_adoption/permissions.py
from rest_framework.permissions import BasePermission, SAFE_METHODS

    
class IsOwner(BasePermission):
    def has_object_permission(self, request, view, obj):
        if obj.user.id == request.user.id:
            return True
        
        return False